requirement.txt
amqp==5.1.1
aniso8601==9.0.1
apispec==5.2.2
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.7
billiard==3.6.4.0
boto3==1.24.7
botocore==1.27.7
cachetools==3.1.1
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
deepdiff==5.8.1
Deprecated==1.2.13
dill==0.3.2
dnslib==0.9.19
dnspython==2.2.1
docutils==0.16
ecdsa==0.17.0
Flask==1.1.2
flask-restplus==0.13.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.0
jsonschema==3.2.0
kombu==5.2.4
localstack-client==1.35
localstack-ext==0.14.3.5
MarkupSafe==1.1.1
ordered-set==4.1.0
packaging==21.3
pbr==5.9.0
plux==1.3.1
prompt-toolkit==3.0.29
psutil==5.9.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.16.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
PyYAML==5.4.1
redis==3.5.3
redis-py-cluster==2.1.0
requests==2.25.1
rich==12.4.4
rsa==4.7.2
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
stevedore==3.5.0
tabulate==0.8.9
tailer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.16.1
wrapt==1.14.1
zipp==3.8.0
